pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[project]
name = "msgraphx"
version = "0.1.0"
description = "Microsoft Graph eXploitation toolkit"
authors = [
    {name = "n3rada",email = "72791564+n3rada@users.noreply.github.com"}
]
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
requires-python = ">=3.9,<4.0"

dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "httpx[socks] (>=0.28.1,<0.29.0)",
    "pkce (>=1.0.3,<2.0.0)",
    "msgraph-sdk (>=1.37.0,<2.0.0)",
    "prompt-toolkit (>=3.0.51,<4.0.0)"
]

[project.urls]
repository = "https://github.com/n3rada/msgraphx"

[project.scripts]
msgraphx = 'msgraphx.cli:main'

